Increased GND path width
Increased the 12V to switch and GND path widths to allow higher current. This can be particularly useful if we are going to add motor drivers and attach motors directly to the board. 
opinion: shud I add external 3V3 or not needed? (I think not)
future work: Add slots for DHT11/ 22